chore: clean up with detekt
